make use of the digest helper and also handle digest on every incoming block if present

+++ src/main/java/com/limechain/network/protocol/grandpa/messages/consensus/GrandpaConsensusMessage.java
@@ -11,6 +11,6 @@ public class GrandpaConsensusMessage {
     private BigInteger delayStartBlockNumber;
     private List<Authority> authorities;
     private BigInteger disabledAuthority;
-    private long delay;
+    private BigInteger delay;
     private GrandpaConsensusMessageFormat format;
 }
